add arrows and comfirm

+++ src/components/BudgetList.vue
@@ -3,10 +3,12 @@
     <h4>Budget List</h4>
     <ul v-if="list.length > 0">
       <li v-for="(item, index) in list" :key="index">
+        <img v-if="item.type === 'INCOME'" src="../assets/up-arrow.svg" alt="">
+        <img v-else src="../assets/download-arrow.svg" alt="">
         <span class="comment">{{ item.comment }} {{ index }}</span>
-        <span :class="[item.type ==='INCOME' ? 'income': 'outcome', 'value']">
+        <span :class="[item.type === 'INCOME' ? 'income' : 'outcome', 'value']">
           {{ item.value }}
-          </span>
+        </span>
         <button @click="deleteItem(item.id)">Delete</button>
       </li>
     </ul>
@@ -20,7 +22,13 @@ export default {
   props: ['list'],
   methods: {
     deleteItem(id) {
-      this.$emit('deleteItem', id);
+      // eslint-disable-next-line no-alert
+      // eslint-disable-next-line no-restricted-globals
+      const deleteConfirm = confirm('Удалить это изменение бюджета?');
+      if (deleteConfirm) {
+        this.$emit('deleteItem', id);
+        // eslint-disable-next-line no-useless-return
+      } else return;
     },
   },
 };
@@ -47,6 +55,12 @@ li {
   align-items: center;
   margin: 0 20px 20px 20px;
 }
+img{
+  width: 14px;
+  margin-right: 5px;
+  stroke: red;
+}
+
 .comment::first-letter {
   text-transform: uppercase;
 }
@@ -54,10 +68,10 @@ li {
   margin: 0 20px 0 auto;
   font-weight: bold;
 }
-.income{
+.income {
   color: green;
 }
-.outcome{
+.outcome {
   color: red;
 }
 button {
